refactor(pieces): rename package from 'pieces' to 'piece'
