+++ app/src/main/java/org/chess/domain/piece/Piece.java
@@ -1,4 +1,4 @@
-package org.chess.domain.pieces;
+package org.chess.domain.piece;
 
 import java.util.Objects;
